+++ .coveragerc
@@ -1,0 +1,20 @@
+[run]
+branch = True
+
+omit =
+    */ENV/*
+    */migrations/*
+    */test/*
+    */admin.py
+    */permissions.py
+    */apps.py
+    */__init__.py
+    */tests.py
+    */manage.py
+    */wsgi.py
+
+source =
+    imagersite
+
+[report]
+show_missing = True
