--- imagersite/imager_images/tests.py
@@ -1,153 +1,0 @@
-from django.test import TestCase, RequestFactory
-from .views import library_view
-from .models import Photo, Album, User
-from imager_profile.tests import UserFactory
-import factory
-from django.urls import reverse_lazy, reverse
-
-
-class PhotoFactory(factory.django.DjangoModelFactory):
-    """Factory that creates fake photos for testing"""
-    class Meta:
-        model = Photo
-    title = factory.Faker('text', max_nb_chars=100)
-    description = factory.Faker('text')
-
-
-class AlbumFactory (factory.django.DjangoModelFactory):
-    """Factory that creates fake albums for testing"""
-    class Meta:
-        model = Album
-    title = factory.Faker('text', max_nb_chars=100)
-    description = factory.Faker('text')
-
-
-class PhotoAlbumUnitTests(TestCase):
-    @classmethod
-    def setUpClass(cls):
-        super(TestCase, cls)
-        for _ in range(5):
-            user = UserFactory.create()
-            # user = User.objects.create(username='testuser')
-            user.set_password('CodeFellows')
-            user.save()
-
-            album = AlbumFactory.create(user=user)
-            album.save()
-
-            photo = PhotoFactory.create(user=user)
-            photo.save()
-
-    @classmethod
-    def tearDownClass(cls):
-        """tears down test objects"""
-        super(TestCase, cls)
-        Photo.objects.all().delete()
-        Album.objects.all().delete()
-
-    def test_photo_is_created(self):
-        """tests if photos were created successfully"""
-        self.assertTrue(Photo.objects.count())
-
-    def test_album_is_created(self):
-        """tests if album were created successfully"""
-        self.assertTrue(Album.objects.count())
-
-    def test_album_has_multiple_photos_and_users(self):
-        """tests if album takes multiple photos owned by multiple users"""
-        album = Album.objects.first()
-        album.save()
-        photo1 = Photo.objects.first()
-        photo2 = Photo.objects.all()[1]
-        photo1.user = User.objects.first()
-        photo2.user = User.objects.all()[1]
-        photo1.save()
-        photo2.save()
-        album.photos.add(photo1, photo2)
-        photo_count = album.photos.count()
-        self.assertTrue(photo_count == 2)
-        self.assertTrue(photo1.user is not photo2.user)
-
-    def test_photo_in_multiple_albums(self):
-        """tests if single photo instance can be in multiple albums"""
-        album1 = Album.objects.all()[2]
-        album2 = Album.objects.all()[3]
-        album1.user = User.objects.all()[2]
-        album2.user = User.objects.all()[2]
-        photo = Photo.objects.all()[2]
-        photo.user = User.objects.all()[2]
-        album1.photos.add(photo)
-        album2.photos.add(photo)
-        album1.save()
-        album2.save()
-        photo.save()
-        self.assertTrue(photo.albums.count() == 2)
-
-
-class ViewTests(TestCase):
-    @classmethod
-    def setUpClass(cls):
-        """Set up some users, albums and photos."""
-        super(TestCase, cls)
-        for _ in range(5):
-            user = UserFactory.create()
-            user.set_password('CodeFellows')
-            user.save()
-
-            album = AlbumFactory.create(user=user)
-            album.save()
-
-            photo = PhotoFactory.create(user=user)
-            photo.save()
-
-    @classmethod
-    def tearDownClass(cls):
-        """tears down test objects"""
-        super(TestCase, cls)
-        Photo.objects.all().delete()
-        Album.objects.all().delete()
-
-    def test_photo_view(self):
-        """Validate photo view exists and renders."""
-        photo = Photo.objects.all().first()
-        photo.user = User.objects.all().first()
-        photo.published = 'PUBLIC'
-        photo.save()
-        response = self.client.get(reverse('photo', args=[photo.id]))
-        self.assertEqual(response.status_code, 200)
-        self.assertTemplateUsed(response, 'imager_images/photo.html')
-
-    def test_album_view(self):
-        """Validate album view exists and renders."""
-        album = Album.objects.all().first()
-        album.user = User.objects.all().first()
-        album.save()
-        response = self.client.get(reverse('album', args=[album.id]))
-        self.assertEqual(response.status_code, 200)
-        self.assertTemplateUsed(response, 'imager_images/album.html')
-
-    def test_photo_gallery_view(self):
-        """Validate photo_gallery view exists and renders."""
-        response = self.client.get(reverse('photo_gallery'))
-        self.assertEqual(response.status_code, 200)
-        self.assertTemplateUsed(response, 'imager_images/photo_gallery.html')
-
-    def test_album_gallery_view(self):
-        """Validate album_gallery view exists and renders."""
-        response = self.client.get(reverse('album_gallery'))
-        self.assertEqual(response.status_code, 200)
-        self.assertTemplateUsed(response, 'imager_images/album_gallery.html')
-
-    def test_library_view_redirect(self):
-        """Validate library view redirects without login."""
-        response = self.client.get(reverse('library'), follow=True)
-        self.assertEqual(response.status_code, 200)
-
-    def test_library_view_user(self):
-        """Validate library view exists and is rendered with user"""
-        user = User.objects.all().first()
-        self.client.login(username=user.username, password=user.password)
-        request = RequestFactory().get(reverse('library'))
-        request.user = user
-        response = library_view(request)
-        self.assertEqual(response.status_code, 200)
