+++ imagersite/imager_profile/tests/test_models.py
@@ -1,7 +1,6 @@
 from django.test import TestCase
-from .models import ImagerProfile, User
+from ..models import ImagerProfile, User
 import factory
-from django.urls import reverse_lazy
 
 
 class UserFactory(factory.django.DjangoModelFactory):
@@ -68,11 +67,3 @@ class ProfileUnitTests(TestCase):
         profile.is_active = False
         profile.save()
         self.assertFalse(profile in ImagerProfile.active())
-
-    def test_profile(self):
-        """testing that profile path renders successful"""
-        user = User.objects.first()
-        ProfileFactory.create(user=user)
-        response = self.client.get(
-            reverse_lazy('named_profile', args=[user.username]))
-        self.assertTrue(response.status_code == 200)
