+++ imagersite/imagersite/views.py
@@ -1,21 +1,16 @@
 from django.shortcuts import render
-from imager_images.models import Photo, Album
+from imager_images.models import Photo
 
 
 def home_view(request):
     '''Default "home" route'''
-    
+
     context = {
-        'title': 'ImagerSite'
+        'title': 'ImagerSite Home'
     }
 
     if Photo.objects.all().count():
-        context['banner'] = Photo.objects.filter(published='PUBLIC').order_by('?').first()
+        context['banner'] = Photo.objects.filter(
+            published='PUBLIC').order_by('?').first()
 
     return render(request, 'home.html', context)
-
-
-# def user_profile_view(request):
-#     '''Default landing page after login.'''
-#     return render(
-#         request, 'user_profile.html', {'title': 'ImagerSite > User Profile'})
